General and Query MemoryManager implemented (untested)
Only malloc and free supported so far. free( void * ) ends up in a NOP. malloc( size_t ) is dispatched to the thread_local instance of the Query Memory Manager (QMM). The QMM looks up within its own space whether it can serve the request. If so, a pointer is returned and the QMM space is updated. If not, the requested space is *expanded* (to actually allocate a big chunk) and dispatched to the General Memory Manager (GMM). The GMM calls the original malloc and holds all references of allocated chunks (bins).
If the QMM gets destructed, GMM is called with a pointer of the QMM. The GMM takes this pointer to wipe out all memory bins of the specific QMM.